Added Table component/added utils/ added custom hooks
