+++ .eslintrc.json
@@ -8,6 +8,18 @@
         "airbnb",
         "prettier"
     ],
+    "settings": {
+        "import/resolver": {
+            "node": {
+                "extensions": [
+                    ".js",
+                    ".jsx",
+                    ".ts",
+                    ".tsx"
+                ]
+            }
+        }
+    },
     "parser": "@typescript-eslint/parser",
     "parserOptions": {
         "ecmaFeatures": {
@@ -25,6 +37,7 @@
             "error",
             2
         ],
+        "import/prefer-default-export": "off",
         "@typescript-eslint/no-explicit-any": "error",
         "@typescript-eslint/no-unused-vars": "error",
         "@typescript-eslint/no-useless-constructor": "error",
@@ -73,12 +86,12 @@
             "error",
             "ignorePackages",
             {
-                "js": "never",
-                "jsx": "never",
-                "ts": "never",
-                "tsx": "never"
+              "js": "never",
+              "jsx": "never",
+              "ts": "never",
+              "tsx": "never"
             }
-        ],
+         ],
         "react/react-in-jsx-scope": "off"
     }
 }
