+++ src/utils/functions.ts
@@ -1,0 +1,2 @@
+export const numberWithCommas = (number: number) =>
+  number.toString().replace(/\B(?<!\.\d*)(?=(\d{3})+(?!\d))/g, ',');
